+++ modules/Food_Service/ServeMenus.php
@@ -76,6 +76,8 @@ else
 	$_SESSION['FSA_menu_id'] = $_REQUEST['menu_id'];
 }
 
+$menu_title = $menus_RET[$_REQUEST['menu_id']][1]['TITLE'];
+
 if ( $_REQUEST['modfunc'] === 'add' )
 {
 	if ( ! empty( $_REQUEST['item_sn'] ) )